+++ module/Core/src/EventDispatcher/UpdateGeoLiteDb.php
@@ -14,6 +14,7 @@ use Throwable;
 
 use function sprintf;
 
+/** @todo Rename to UpdateGeolocationDb */
 readonly class UpdateGeoLiteDb
 {
     public function __construct(